# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample recipes API. You can find out more about the API at https://github.com/XupyprY/srv_recipes.git",
    "title": "Recipes API",
    "contact": {
      "name": "Yuri Radionov",
      "email": "45962539+XupyprY@users.noreply.github.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/auth/logout": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Logout: удаляет refresh токен из Redis, требует Authorization заголовок с Bearer токеном",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Successful logout"
          },
          "401": {
            "description": "Unauthorized (token missing or invalid)"
          }
        }
      }
    },
    "/auth/me": {
      "post": {
        "description": "Me: сохраняет и возвращает сессию",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current session information",
            "schema": {
              "type": "object",
              "properties": {
                "ip": {
                  "type": "string"
                },
                "location": {
                  "type": "string"
                },
                "userAgent": {
                  "type": "string"
                },
                "userID": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (token missing or invalid)"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Refresh: проверяет refresh в Redis, выдаёт новую пару",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Token is new and doesn't need a refresh"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "description": "Login: возвращает access+refresh, сохраняет refresh в Redis",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "signIn",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "description": "Принимает email и пароль для регистрации пользователя.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Создает нового пользователя",
        "operationId": "signUp",
        "parameters": [
          {
            "description": "Email и пароль для регистрации",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "strongpassword123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь успешно создан",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string",
                  "example": "short_life_access_token"
                },
                "refresh_token": {
                  "type": "string",
                  "example": "long_life_refresh_token"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос (например, отсутствуют поля email или password)"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "description": "Verify: принимает verify token, делает verified=true",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "verify",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Verification error"
          }
        }
      }
    },
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "listRecipes"
      },
      "post": {
        "description": "Create a new recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "newRecipe",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "CustomID",
            "name": "customId",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Tags",
            "name": "tags",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Ingredients",
            "name": "ingredients",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Instructions",
            "name": "instructions",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "PublishedAt",
            "name": "publishedAt",
            "in": "query"
          }
        ]
      }
    },
    "/recipes/search": {
      "get": {
        "description": "Search recipes based on tags",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "findRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "recipe tag",
            "name": "tag",
            "in": "query",
            "required": true
          }
        ]
      }
    },
    "/recipes/{id}": {
      "get": {
        "description": "Get one recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "oneRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "description": "Update an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "updateRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "description": "Delete an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      }
    }
  },
  "definitions": {
    "SignUpRequest": {
      "description": "SignUpRequest represents the request body for user signup",
      "type": "object",
      "required": [
        "Email",
        "Password"
      ],
      "properties": {
        "Email": {
          "description": "User's email address",
          "type": "string",
          "example": "user@example.com"
        },
        "Password": {
          "description": "User's password",
          "type": "string",
          "example": "strongpassword"
        }
      },
      "x-go-package": "srv_recipes/handlers"
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}